// src/app/core/models/ILead.interface.ts
export interface ILead {
  id: string;
  name: string;
  address: string;
  position: string;
  city: string;
  email: string;
  province: string;
  website: string;
  country: string;
  phone: string;
  zipCode: string;
  defaultLanguage: string;
  company: string;
  description: string;
  status: string;
  source: string;
  assignedTo: string;
}